Fix bug causing CI to fail due to pushing Docker images without permission
For external PRs, there is no permission to upload to GHCR.io.  Attempting to do so merely causes the CI to always fail.  This PR fixes that issue so it won't attempt to if the triggering pull request is from an *external* repository, which means the CI will not have permission to push to GHCR.